+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -11,6 +11,7 @@ however, it has to be formatted properly to pass verification tests.
 ## [Unreleased] - yyyy-mm-dd
 
 ### Fixed
+- Fixed an issue where removing a newly created action in the Asset Editor would cause an exception. [UUM-95693](https://issuetracker.unity3d.com/product/unity/issues/guid/UUM-95693)
 - Fixed gamepad navigation in UI Toolkit TextField when using InputSystemUIInputModule.
 
 ## [1.13.0] - 2025-02-05